requirements.txt
-i https://pypi.org/simple
amqp==2.6.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
arabic-reshaper==2.1.1
asgiref==3.3.1; python_version >= '3.5'
attrs==20.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
awesome-slugify==1.6.5
bcrypt==3.2.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==4.4.6
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cryptography==3.4.6; python_version >= '3.6'
dicttoxml==1.7.4
django-admin-autocomplete-filter==0.6.1
django-ckeditor==6.0.0
django-cors-headers==3.5.0
django-defender==0.8.0
django-extra-fields==3.0.2
django-filter==2.4.0
django-grappelli==2.14.3
django-ipware==3.0.2
django-js-asset==1.2.2
django-modeltranslation==0.16.1
django-recaptcha==2.0.6
django-redis==4.12.1
django-rest-knox==4.1.0
django-rest-resetpassword==0.1.2
django-rosetta==0.9.5
django-upload-validator==1.1.5
django==3.1.6
djangorestframework-camel-case==1.2.0
djangorestframework==3.12.2
drf-nested-forms==0.2.1
drf-recaptcha==2.0.4
extruct==0.11.0
factory-boy==3.2.0; python_version >= '3.6'
fake-useragent==0.1.11
faker==6.3.0; python_version >= '3.6'
future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
html-text==0.5.2
html5lib==1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
inflection==0.5.1; python_version >= '3.5'
iniconfig==1.1.1
isodate==0.6.0
jstyleson==0.0.2
kombu==4.6.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
lxml==4.6.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
mf2py==1.1.2
packaging==20.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
passlib==1.7.4
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0
pillow==8.0.1
pluggy==0.13.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
polib==1.1.0
psycopg2-binary==2.8.6
py==1.10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyopenssl==20.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pypdf2==1.26.0
pytest-django==4.1.0
pytest-factoryboy==2.0.3
pytest==6.2.2; python_version >= '3.6'
python-bidi==0.4.2
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-magic==0.4.22; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pytz==2021.1
rdflib-jsonld==0.5.0
rdflib==4.2.2
redis==3.5.3
regex==2020.11.13
reportlab==3.5.56
requests==2.25.1
sentry-sdk==0.20.2
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.2; python_version >= '3.0'
sqlparse==0.4.1; python_version >= '3.5'
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
unidecode==0.4.21
urllib3==1.26.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
uwsgi==2.0.19.1
vine==1.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
w3lib==1.22.0
webencodings==0.5.1
xhtml2pdf==0.2.5
xlsxwriter==1.3.7
